Allergy And Dislikes

+++ DL/DtosV1/Allergies/PutAllergyDto.cs
@@ -4,6 +4,6 @@ namespace DL.DtosV1.Allergies
 {
     public class PutAllergyDto
     {
-        public HashSet<int> AllergyIds { get; set; }
+        public HashSet<long> AllergyIds { get; set; }
     }
 }
